auth

+++ src/main/java/com/example/url_monitor/Monitor/MonitorService.java
@@ -7,6 +7,8 @@ import com.example.url_monitor.User.UserEntity;
 import com.example.url_monitor.User.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,6 +23,13 @@ public class MonitorService {
     private final CheckMonitorService monitor_check_service;
     private final LogService log_service;
     public MonitorEntity CreateMonitor(MonitorDTO request){
+
+        if (request.getMonitor_type() == MonitorType.HTTP) {
+            String url = request.getMonitor_url();
+            if (!url.startsWith("http://") && !url.startsWith("https://")) {
+                throw new RuntimeException("HTTP monitor URL must start with http:// or https://");
+            }
+        }
         MonitorEntity monitor = MonitorEntity.builder()
                 .nameVar(request.getMonitor_name())
                 .urlVar(request.getMonitor_url())
@@ -34,27 +43,52 @@ public class MonitorService {
                 .requestBodyVar(request.getRequest_body())
                 .requestHeadersVar(request.getRequest_headers())
                 .build();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity user = (UserEntity) authentication.getPrincipal();
 
-        UserEntity user = user_repository.findById(1L).orElseThrow(() -> new RuntimeException("User not found"));
         monitor.setUserVar(user);
         monitor = monitor_repository.save(monitor);
         log_service.CreateCreateLog(monitor,user);
         return monitor;
     }
     public MonitorEntity GetMonitor(Long id){
-        return monitor_repository.findById(id).orElseThrow(() -> new RuntimeException("Monitor not found"));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity user = (UserEntity) authentication.getPrincipal();
+
+        MonitorEntity monitor = monitor_repository.findById(id).orElseThrow(() -> new RuntimeException("Monitor not found"));
+
+        if (!monitor.getUserVar().getId().equals(user.getId())) throw new RuntimeException("Forbidden");
+
+        return monitor;
     }
     public List<MonitorEntity> GetAllMonitors(){
-        return monitor_repository.findAll();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity user = (UserEntity) authentication.getPrincipal();
+        return monitor_repository.findAllByUserVar(user);
     }
     public void DeleteMonitor(Long id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity user = (UserEntity) authentication.getPrincipal();
         MonitorEntity monitor = GetMonitor(id);
+        if (!monitor.getUserVar().getId().equals(user.getId())) throw new RuntimeException("Forbidden");
         log_repository.deleteAllByMonitorVar(monitor);
         monitor_repository.deleteById(id);
     }
 
     public MonitorEntity UpdateMonitor(Long id, MonitorDTO request){
+
+        if (request.getMonitor_type() == MonitorType.HTTP) {
+            String url = request.getMonitor_url();
+            if (!url.startsWith("http://") && !url.startsWith("https://")) {
+                throw new RuntimeException("HTTP monitor URL must start with http:// or https://");
+            }
+        }
         MonitorEntity monitor=monitor_repository.findById(id).orElseThrow(() -> new RuntimeException("Monitor not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity user = (UserEntity) authentication.getPrincipal();
+        if (!monitor.getUserVar().getId().equals(user.getId())) throw new RuntimeException("Forbidden");
+
         monitor.setNameVar(request.getMonitor_name());
         monitor.setUrlVar(request.getMonitor_url());
         monitor.setMonitor_type_var(request.getMonitor_type());
